Implement user_friendship model and fix model associations setup
This update will provide the user the queries and mutations for
sending and responding to friend requests. Also, there was a bug where
model associations didn't work because the module for setting up
associations was not imported within app startup. Now, the module will
be imported within app startup, and model associations work properly