Implement error formatter for the apollo server instance
The application's apollo server instance used to respond to the client
with verbose errors, which may distract the client in production. There
will now be a error formatter implemented to simplify the response
error's structure. It will also store the error's stack trace for
development purposes in non-production environments only