Fix migrations to set timestamp precisions to milliseconds
Migrations used to set timestamps to a precision of microseconds, which
prevented queries of timestamp = date from succeeding because JS Date
only works down to milliseconds. This is problematic because when
querying timestamp attributes with cursor-based pagination, it is
required to handle rows with equal timestamps. Timestamp columns will
now be set to a precision of milliseconds