Implement production-ready event-publishing system for subscriptions
The event-publishing system for GraphQL subscriptions used to take
place in memory. This is problematic because events published by one
GraphQL server instance won't be received by subscriptions that are
handled by other instances. Events will now be published in the Postgres
database, so that all servers can listen to published events