Implement display name and fix profile picture upload
The name shown in chats and on friends and friend requests will now be
the user's display name. The display name will be username#discriminator
by default, and Client will be able to edit this in their profile
settings page. Profile picture upload was broken due to backend syntax
errors, which will now be fixed