Implement frontend layout refactor and sidebar collapse buttons
The layout for pages on the frontend will now rely on flex so that
the layout adapts easily to collapsed elements. Client will now be able
to click on a collapse button on the sidebar to collapse it. When they
bring their cursor near the collapse sidebar, the collapse button will
reappear for them to open the sidebar