Implement fix for user avatar's online/offline status and user page
The online/offline indicator on a user's avatar used to disappear after
a custom avatar picture was set. The indicator will now remain after the
a new avatar is set. Also, the user page used to compress to the side
due to not having flex grow active. The user page will now size
responsively