Implement fix for chat search and remove mongo transactions
The chat search used to search chats by name, but chats by default have
a null name because a chat's name is determined on the client side using
the names of the users in the chat. The chat search will now search
chats by the names of its users. Mongo transactions require mongodb
replica sets which cost money and so will be disabled