Implement backend startup controller
When the backend receives no traffic for a while, Cloud Run scales its
container instances to 0. When the frontend requests the backend at 0
instances, the backend scales to 1 instance, which causes a server to
start up. During start up, the server takes time to connect to the
databases before it sets up its API GraphQL route. This is a problem
because, during this time, requests are responded to with 404.

The backend will now attach a middleware during startup that handles all
network requests by responding "Server starting up, please try again in
a few seconds." The frontend's error handling will be updated to
handle this non-GraphQL response in its error components.